moved command function
